requirements/base.txt
# Django Core
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment
python-dotenv==1.0.1
dj-database-url==2.2.0

# Caching & Sessions
django-redis==5.4.0
redis==5.0.8

# Celery & Background Tasks
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# WebSockets
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# ASGI Server
uvicorn[standard]==0.30.6